# swagger/api.yaml
openapi: 3.0.0
info:
  description: "This is the static API documentation for the project called medical-metadata which deals as a proof of concept for the master thesis of Bence Slajchó @Alpen-Adria-University-Klagenfurt. Supervisors: O.Univ.-Prof. Dipl.-Ing. Dr. Johann Eder; Dr.phil. Volodymyr Shekhovtsov"
  version: "1.3.0"
  title: "Medical Metadata"
  contact:
    email: "slajchobence@gmail.com"
...
servers:
  - url: localhost:3001
    description: development server
paths:
...
  /api/metadata/metadimension:
    put:
      tags:
      - "metadata interface"
      summary: "loads the metadata into the system alongside custom principles"
      description: "the metadata config objects are neccessary to be found in a state which can be read by the system"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/metadataLoadingApi"
      responses:
        "200":
          description: "successful request"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
    delete:
      tags:
      - "metadata interface"
      summary: delete metadata for a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                collectionName:
                  type: "string"
                  example: "raw_medical_data_1"
      responses:
        "200":
          description: "successful request"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # metadata import from external file
  /api/metadata/export:
    post:
      tags:
      - "metadata interface"
      summary: "creates a JSON structured export file which includes metadata"
      description: "The metadata gets created alongside the parameters which the user passes to this endpoint. This data is the same which also gets accepted which the system is using for metadata creation directly into the system."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/metadataLoadingApi"
      responses:
        "200":
          description: "successful request"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # metadata import
  /api/metadata/import:
    post:
      tags:
      - "metadata interface"
      summary: "loads collection metadata into the system which was created by an external source"
      description: To offer a possibility to load directly metadata into the system (and not sharing collection data of any kind with the application) the metadata import interface was introduced. 
        It accepts a JSON file which contains all the needed metadata information which otherwise the system would create.
        The structure of the JSON file is described in detail in the thesis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: "successful request"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  ### 
  /api/table-cols-for-scaling:
    get:
      tags:
      - "aggregator"
      summary: "returns a set of a certain scaled table columns"
      parameters:
        - in: query
          name: scaling
          required: true
          schema:
            type: string
            example: "nom"
        - in: query
          name: table_name
          required: true
          schema:
            type: string
            example: "collection_one"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTableColsForScaling"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
  /api/collection-list:
    get:
      tags:
      - "aggregator"
      summary: "list of collections"
      description: "not aggregated yet"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCollectionList"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
  /api/distinct-values:
    get:
      tags:
      - "aggregator"
      summary: "returns information of available ordinal values for the loinc code"
      description: "For Ord scaled LOINC code the available values are returned"
      parameters:
        - in: query
          name: column_name
          required: true
          schema:
            type: string
            example: "ur_bld"
        - in: query
          name: table_name
          required: true
          schema:
            type: string
            example: "collection_one"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOrdinalDistinctValues"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
  /api/ordinal-valuesets-for-loinc:
    get:
      tags:
      - "aggregator"
      summary: "list of available ordinal value sets"
      parameters:
        - in: query
          name: loincCode
          required: true
          schema:
            type: string
            example: "5794-3"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOrdinalValueSets"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
  /api/nominal-mapped-value:
    get:
      tags:
      - "aggregator"
      summary: "list of existing values to which the mapping can be done of a certain loinc"
      parameters:
        - in: query
          name: loincCode
          required: true
          schema:
            type: string
            example: "5778-6"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOrdinalNominalMappedValue"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
  /api/ordinal-valuesets-values:
    get:
      tags:
      - "aggregator"
      summary: "list of existing values of a certain value set for ordinal mapping"
      parameters:
        - in: query
          name: loincCode
          required: true
          schema:
            type: string
            example: "5778-6"
        - in: query
          name: valueSetName
          required: true
          schema:
            type: string
            example: "numeric scale"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOrdinalOrdinalValueSetValues"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
...
  /api/metadata/quality-completeness:
    get:
      tags:
      - "metadata quality"
      summary: "Calculates the completeness of a a LOINC code which belong to a collection as a descriptor"
      parameters:
        - in: query
          name: loincCode
          required: true
          schema:
            type: string
            example: "2161-8"
        - in: query
          name: lowerBound
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - in: query
          name: upperBound
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 100
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseQualityCompleteness"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
  /api/metadata/quality-accuracy:
    get:
      tags:
      - "metadata quality"
      summary: "Calculates the completeness of a a LOINC code which belong to a collection as a descriptor"
      parameters:
        - in: query
          name: loincCode
          required: true
          schema:
            type: string
            example: "2161-8"
        - in: query
          name: accuracy
          required: true
          schema:
            type: array
            items:
              type: string
              example: ['low']
              enum: ['low', 'medium', 'high']
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseQualityAccuracy"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
  /api/metadata/quality-timeliness:
    get:
      tags:
      - "metadata quality"
      summary: "Calculates the completeness of a a LOINC code which belong to a collection as a descriptor"
      parameters:
        - in: query
          name: loincCode
          required: true
          schema:
            type: string
            example: "2161-8"
        - in: query
          name: lowerBound
          required: true
          schema:
            type: integer
            example: 0
        - in: query
          name: upperBound
          required: true
          schema:
            type: integer
            example: 365
        - in: query
          name: timelinessUnit
          required: true
          schema:
            type: string
            example: "day"
            enum: ["day", "month", "year"]
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseQualityTimeliness"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
components:
  schemas:
...
    ApiResponseError:
      type: "object"
      properties:
        status:
          type: "integer"
          format: "123"
          enum: [500]
        message:
          type: "string"
...
    metadataLoadingApi:
      type: object
      properties:
        collectionName:
          type: string
          example: 'customCollectionOne'
        accuracySupport:
          type: string
          enum: ['high', 'medium', 'low']
        ordinalAsNominal:
          type: boolean
        narrativeSupport:
          type: boolean
        documentSupport:
          type: boolean
        ordinalMapperObject:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: collection1
                description: 'First level of the tree - collection'
              name:
                type: string
                example: Urinary sample collection
              children:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "5794-3"
                      description: 'Second level of the tree - loinc'
                    name:
                      type: string
                      example: "5794-3"
                    children:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: string
                            description: 'Third level of the tree - value (origin)'
                            example: '1+'
                          name:
                            type: string
                            description: 'Third level of the tree - value (mapped onto)'
                            example: '1'
        nominalMappingConsidered:
          type: boolean
        nominalMappingObject:
          type: object
          properties:
            collectionName:
              description: the name of the property is the name of the collection
              type: array
              items: 
                type: object
                properties:
                  items:
                    type: object
                    properties:
                      loincCode: 
                        description: the name of the property is the loinc code
                        type: array
                        items:
                          type: object
                          properties:
                            aggregateValue:
                              type: string
                              example: "amber"
                            selectedValue:
                              type: string
                              example: "dark"
    ApiResponseCollectionList:
      type: "array"
      items:
        type: object
        properties:
          hasindexes:
            type: boolean
          hasrules:
            type: boolean
          hastriggers:
            type: boolean
          rowsecurity:
            type: boolean
          schemaname:
            type: string
            example: "public"
          tablename:
            type: string
            example: "collection_one"
          tableowner:
            type: string
            example: "postgres"
    ApiResponseTableColsForScaling:
      type: "array"
      items:
        type: object
        properties:
          column_name:
            type: string
            example: "ur_color"
          maps_to_loinc:
            type: string
            example: "5778-6"
          table_name:
            type: string
            example: "collection_one"
    ApiResponseOrdinalDistinctValues:
      type: "array"
      items:
        type: object
        properties:
          aggregate_value:
            type: string
            example: "0"
          column_name:
            type: string
            example: "ur_bld"
          loinc_code:
            type: string
            example: "5794-3"
          occurance_count:
            type: number
            example: 9
          source_database:
            type: string
            example: "collection_one"
    ApiResponseOrdinalValueSets:
      type: "array"
      items:
        type: object
        properties:
          value_set_name:
            type: string
            example: "numeric scale"
    ApiResponseOrdinalNominalMappedValue:
      type: "array"
      items:
        type: object
        properties:
          nominal_valiue:
            type: string
    ApiResponseOrdinalOrdinalValueSetValues:
      type: "array"
      items:
        type: object
        properties:
          loinc_code:
            type: string
            example: "5794-3"
          sequence_number:
            type: number
            example: 1
          value_set_name:
            type: string
            example: "numeric scale"
          value_set_value:
            type: string
            example: "1"
...
    ApiResponseQualityCompleteness:
      type: array
      items:
        type: object
        properties:
          source_database: 
            type: string
            example: "raw_medical_data_1"
          collection_completeness: 
            type: number
            example: "53.93439343934393439300"
          null_value_count: 
            type: number
            example: "4006"
          occurance_count: 
            type: number
            example: "7748"
          standard_metadata_completeness: 
            type: number
            example: "48.48913462774848913500"
    ApiResponseQualityAccuracy:
      type: array
      items:
        type: object
        properties:
          collection_accuracy: 
            type: number
            example: "0.8055144586415602"
          source_database: 
            type: string
            example: "raw_medical_data_1"
          occurance_count: 
            type: number
            example: "160"
          tuple_accuracy: 
            type: string
            example: "low"
    ApiResponseQualityTimeliness:
      type: array
      items:
        type: object
        properties:
          collection_timeliness: 
            type: number
            example: 32.22394010559027
          source_database: 
            type: string
            example: "raw_medical_data_1"
          occurance_count: 
            type: number
            example: "1111"
          timeliness: 
            type: number
            example: 125
...